refactor: change biginteger to double type

+++ src/main/java/com/carrental/dto/ExtraFeeDto.java
@@ -2,13 +2,11 @@ package com.carrental.dto;
 
 import lombok.Data;
 
-import java.math.BigInteger;
-
 @Data
 public class ExtraFeeDto {
     private Long id;
     private String name;
     private Long limit;
     private String unit;
-    private BigInteger fee;
+    private Double fee;
 }
