+++ src/main/java/com/carrental/requestmodel/ServiceFeeRequest.java
@@ -2,12 +2,10 @@ package com.carrental.requestmodel;
 
 import lombok.Data;
 
-import java.math.BigInteger;
-
 @Data
 public class ServiceFeeRequest {
     private Long serviceTypeId;
-    private BigInteger defaultPrice;
+    private Double defaultPrice;
     private Integer discountByWeek;
     private Integer discountByMonth;
 }
