+++ src/main/java/com/carrental/service/impl/ExtraFeeService.java
@@ -4,10 +4,8 @@ import com.carrental.entity.ExtraFeeEntity;
 import com.carrental.repository.IExtraFeeRepository;
 import com.carrental.service.IExtraFeeService;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Service;
 
-import java.math.BigInteger;
 import java.util.List;
 
 @Service
@@ -21,9 +19,9 @@ public class ExtraFeeService implements IExtraFeeService {
     }
 
     @Override
-    public BigInteger findDeliveryToTenantFee(Long id) {
+    public Double findDeliveryToTenantFee(Long id) {
         ExtraFeeEntity fee = repository.findFirstByServiceFeeIdAndName(id, "Giao xe tận nơi");
-        if (null == fee) return new BigInteger(String.valueOf(0));
+        if (null == fee) return new Double(String.valueOf(0));
         return fee.getFee();
     }
